requirements.txt
APScheduler==3.6.3
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
cryptography==40.0.1
decorator==5.1.1
Django==4.2
djangorestframework==3.14.0
Pillow==9.5.0
pycparser==2.21
python-telegram-bot==13.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.12
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2023.3
tzlocal==4.3
